@@ -10,7 +10,7 @@ build-backend = "hatchling.build"
 name = "voila"
 description = "Voilà turns Jupyter notebooks into standalone web applications"
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 authors = [
     { name = "Voila Development Team" },
 ]
