@@ -6,64 +6,83 @@
 #                                                                           #
 # The full license is in the file LICENSE, distributed with this software.  #
 #############################################################################
+import errno
 import gettext
-import io
-import sys
 import json
 import logging
-import threading
-import tempfile
 import os
+import random
 import shutil
 import signal
 import socket
+import sys
+import tempfile
+import threading
 import webbrowser
-import errno
-import random
+
+from .voila_identity_provider import VoilaLoginHandler
 
 try:
     from urllib.parse import urljoin
     from urllib.request import pathname2url
 except ImportError:
     from urllib import pathname2url
+
     from urlparse import urljoin
 
 import jinja2
-
 import tornado.ioloop
 import tornado.web
-
-from traitlets.config.application import Application
-from traitlets.config.loader import Config
-from traitlets import Unicode, Integer, Bool, Dict, List, Callable, default
-
-from jupyter_server.services.kernels.handlers import KernelHandler, ZMQChannelsHandler
-from jupyter_server.services.contents.largefilemanager import LargeFileManager
+from jupyter_client.kernelspec import KernelSpecManager
+from jupyter_core.paths import jupyter_config_path, jupyter_path
 from jupyter_server.base.handlers import FileFindHandler, path_regex
 from jupyter_server.config_manager import recursive_update
-from jupyter_server.utils import url_path_join, run_sync
-from jupyter_server.services.config import ConfigManager
+from jupyter_server.services.config.manager import ConfigManager
+from jupyter_server.services.contents.largefilemanager import LargeFileManager
+from jupyter_server.services.kernels.handlers import KernelHandler
+from jupyter_server.services.kernels.websocket import KernelWebsocketHandler
+from jupyter_server.auth.authorizer import AllowAllAuthorizer, Authorizer
+from jupyter_server.auth.identity import PasswordIdentityProvider
+from jupyter_server import DEFAULT_TEMPLATE_PATH_LIST, DEFAULT_STATIC_FILES_PATH
+from jupyter_server.services.kernels.connection.base import (
+    BaseKernelWebsocketConnection,
+)
+from jupyter_server.services.kernels.connection.channels import (
+    ZMQChannelsWebsocketConnection,
+)
+from jupyter_server.auth.identity import (
+    IdentityProvider,
+)
+from jupyter_server.utils import url_path_join
+from jupyter_core.utils import run_sync
 
 from jupyterlab_server.themes_handler import ThemesHandler
 
-from jupyter_client.kernelspec import KernelSpecManager
 
-from jupyter_core.paths import jupyter_config_path, jupyter_path
+from traitlets import Bool, Callable, Dict, Integer, List, Unicode, default, Type, Bytes
+from traitlets.config.application import Application
+from traitlets.config.loader import Config
 
-from .paths import ROOT, STATIC_ROOT, collect_template_paths, collect_static_paths
-from .handler import VoilaHandler
-from .treehandler import VoilaTreeHandler
 from ._version import __version__
-from .static_file_handler import MultiStaticFileHandler, TemplateStaticFileHandler, WhiteListFileHandler
 from .configuration import VoilaConfiguration
 from .execute import VoilaExecutor
 from .exporter import VoilaExporter
-from .shutdown_kernel_handler import VoilaShutdownKernelHandler
-from .voila_kernel_manager import voila_kernel_manager_factory
+from .paths import ROOT, STATIC_ROOT, collect_static_paths, collect_template_paths
 from .request_info_handler import RequestInfoSocketHandler
-from .utils import create_include_assets_functions
+from .shutdown_kernel_handler import VoilaShutdownKernelHandler
+from .static_file_handler import (
+    MultiStaticFileHandler,
+    TemplateStaticFileHandler,
+    WhiteListFileHandler,
+)
+from .tornado.handler import TornadoVoilaHandler
+from .tornado.treehandler import TornadoVoilaTreeHandler
+from .utils import create_include_assets_functions, get_data_dir, pjoin
+from .voila_kernel_manager import voila_kernel_manager_factory
 
 _kernel_id_regex = r"(?P<kernel_id>\w+-\w+-\w+-\w+-\w+)"
+
+HERE = os.path.dirname(os.path.abspath(__file__))
 
 
 def _(x):
@@ -71,25 +90,36 @@ def _(x):
 
 
 class Voila(Application):
-    name = 'voila'
+    name = "voila"
     version = __version__
-    examples = 'voila example.ipynb --port 8888'
+    examples = "voila example.ipynb --port 8888"
 
     flags = {
-        'debug': (
+        "debug": (
             {
-                'Voila': {'log_level': logging.DEBUG},
-                'VoilaConfiguration': {'show_tracebacks': True},
+                "Voila": {"log_level": logging.DEBUG},
+                "VoilaConfiguration": {"show_tracebacks": True},
             },
-            _("Set the log level to logging.DEBUG, and show exception tracebacks in output.")
+            _(
+                "Set the log level to logging.DEBUG, and show exception tracebacks in output."
+            ),
         ),
-        'no-browser': ({'Voila': {'open_browser': False}}, _('Don\'t open the notebook in a browser after startup.')),
-        'show-margins': (
+        "no-browser": (
+            {"Voila": {"open_browser": False}},
+            _("Don't open the notebook in a browser after startup."),
+        ),
+        "show-margins": (
             {
-                'VoilaConfiguration': {'show_margins': True},
+                "VoilaConfiguration": {"show_margins": True},
             },
-            _('Show left and right margins for the "lab" template, this gives a "classic" template look')
-        )
+            _(
+                'Show left and right margins for the "lab" template, this gives a "classic" template look'
+            ),
+        ),
+        "token": (
+            {"Voila": {"auto_token": True}},
+            _(""),
+        ),
     }
 
     description = Unicode(
@@ -105,65 +135,54 @@ class Voila(Application):
         """
     )
     notebook_filename = Unicode()
-    port = Integer(
-        8866,
-        config=True,
-        help=_(
-            'Port of the Voilà server. Default 8866.'
-        )
-    )
+    port = Integer(8866, config=True, help=_("Port of the Voilà server. Default 8866."))
     autoreload = Bool(
         False,
         config=True,
         help=_(
-            'Will autoreload to server and the page when a template, js file or Python code changes'
-        )
+            "Will autoreload to server and the page when a template, js file or Python code changes"
+        ),
     )
-    root_dir = Unicode(config=True, help=_('The directory to use for notebooks.'))
+    root_dir = Unicode(config=True, help=_("The directory to use for notebooks."))
     static_root = Unicode(
         STATIC_ROOT,
         config=True,
-        help=_(
-            'Directory holding static assets (HTML, JS and CSS files).'
-        )
+        help=_("Directory holding static assets (HTML, JS and CSS files)."),
     )
     aliases = {
-        'autoreload': 'Voila.autoreload',
-        'base_url': 'Voila.base_url',
-        'port': 'Voila.port',
-        'static': 'Voila.static_root',
-        'server_url': 'Voila.server_url',
-        'pool_size': 'VoilaConfiguration.default_pool_size',
-        'show_tracebacks': 'VoilaConfiguration.show_tracebacks',
-        'preheat_kernel': 'VoilaConfiguration.preheat_kernel',
-        'strip_sources': 'VoilaConfiguration.strip_sources',
-        'template': 'VoilaConfiguration.template',
-        'theme': 'VoilaConfiguration.theme'
+        "autoreload": "Voila.autoreload",
+        "base_url": "Voila.base_url",
+        "port": "Voila.port",
+        "static": "Voila.static_root",
+        "server_url": "Voila.server_url",
+        "token": "Voila.token",
+        "pool_size": "VoilaConfiguration.default_pool_size",
+        "show_tracebacks": "VoilaConfiguration.show_tracebacks",
+        "preheat_kernel": "VoilaConfiguration.preheat_kernel",
+        "strip_sources": "VoilaConfiguration.strip_sources",
+        "template": "VoilaConfiguration.template",
+        "theme": "VoilaConfiguration.theme",
     }
-    classes = [
-        VoilaConfiguration,
-        VoilaExecutor,
-        VoilaExporter
-    ]
+    classes = [VoilaConfiguration, VoilaExecutor, VoilaExporter]
     connection_dir_root = Unicode(
         config=True,
         help=_(
-            'Location of temporary connection files. Defaults '
-            'to system `tempfile.gettempdir()` value.'
-        )
+            "Location of temporary connection files. Defaults "
+            "to system `tempfile.gettempdir()` value."
+        ),
     )
     connection_dir = Unicode()
 
     base_url = Unicode(
-        '/',
+        "/",
         config=True,
         help=_(
-            'Path for Voilà API calls. If server_url is unset, this will be \
+            "Path for Voilà API calls. If server_url is unset, this will be \
             used for both the base route of the server and the client. \
             If server_url is set, the server will server the routes prefixed \
             by server_url, while the client will prefix by base_url (this is \
-            useful in reverse proxies).'
-        )
+            useful in reverse proxies)."
+        ),
     )
 
     server_url = Unicode(
@@ -171,60 +190,64 @@ class Voila(Application):
         config=True,
         allow_none=True,
         help=_(
-            'Path to prefix to Voilà API handlers. Leave unset to default to base_url'
-        )
+            "Path to prefix to Voilà API handlers. Leave unset to default to base_url"
+        ),
     )
 
     notebook_path = Unicode(
         None,
         config=True,
         allow_none=True,
-        help=_(
-            'path to notebook to serve with Voilà'
-        )
+        help=_("path to notebook to serve with Voilà"),
     )
 
-    template_paths = List(
-        [],
+    template_paths = List([], config=True, help=_("path to jinja2 templates"))
+
+    static_paths = List([STATIC_ROOT], config=True, help=_("paths to static assets"))
+
+    port_retries = Integer(
+        50,
         config=True,
         help=_(
-            'path to jinja2 templates'
-        )
+            "The number of additional ports to try if the specified port is not available."
+        ),
     )
 
-    static_paths = List(
-        [STATIC_ROOT],
+    ip = Unicode(
+        "localhost",
+        config=True,
+        help=_("The IP address the notebook server will listen on."),
+    )
+
+    open_browser = Bool(
+        True,
         config=True,
         help=_(
-            'paths to static assets'
-        )
-    )
-
-    port_retries = Integer(50, config=True,
-                           help=_("The number of additional ports to try if the specified port is not available.")
-                           )
-
-    ip = Unicode('localhost', config=True,
-                 help=_("The IP address the notebook server will listen on."))
-
-    open_browser = Bool(True, config=True,
-                        help=_("""Whether to open in a browser after starting.
+            """Whether to open in a browser after starting.
                         The specific browser used is platform dependent and
                         determined by the python standard library `webbrowser`
                         module, unless it is overridden using the --browser
                         (NotebookApp.browser) configuration option.
-                        """))
+                        """
+        ),
+    )
 
-    browser = Unicode(u'', config=True,
-                      help="""Specify what command to use to invoke a web
+    browser = Unicode(
+        "",
+        config=True,
+        help="""Specify what command to use to invoke a web
                       browser when opening the notebook. If not specified, the
                       default browser will be determined by the `webbrowser`
                       standard library module, which allows setting of the
                       BROWSER environment variable to override it.
-                      """)
+                      """,
+    )
 
-    webbrowser_open_new = Integer(2, config=True,
-                                  help=_("""Specify Where to open the notebook on startup. This is the
+    webbrowser_open_new = Integer(
+        2,
+        config=True,
+        help=_(
+            """Specify Where to open the notebook on startup. This is the
                                   `new` argument passed to the standard library method `webbrowser.open`.
                                   The behaviour is not guaranteed, but depends on browser support. Valid
                                   values are:
@@ -232,17 +255,24 @@ class Voila(Application):
                                   - 1 opens a new window,
                                   - 0 opens in an existing window.
                                   See the `webbrowser.open` documentation for details.
-                                  """))
+                                  """
+        ),
+    )
 
-    custom_display_url = Unicode(u'', config=True,
-                                 help=_("""Override URL shown to users.
+    custom_display_url = Unicode(
+        "",
+        config=True,
+        help=_(
+            """Override URL shown to users.
                                  Replace actual URL, including protocol, address, port and base URL,
                                  with the given value when displaying URL to the users. Do not change
                                  the actual connection URL. If authentication token is enabled, the
                                  token is added to the custom URL automatically.
                                  This option is intended to be used when the URL to display to the user
                                  cannot be determined reliably by the Jupyter notebook server (proxified
-                                 or containerized setups for example)."""))
+                                 or containerized setups for example)."""
+        ),
+    )
 
     prelaunch_hook = Callable(
         default_value=None,
@@ -268,73 +298,127 @@ class Voila(Application):
         ),
     )
 
+    cookie_secret = Bytes(
+        b"",
+        config=True,
+        help="""The random bytes used to secure cookies.
+        By default this is a new random number every time you start the server.
+        Set it to a value in a config file to enable logins to persist across server sessions.
+
+        Note: Cookie secrets should be kept private, do not share config files with
+        cookie_secret stored in plaintext (you can read the value from a file).
+        """,
+    )
+
+    token = Unicode(None, help="""Token for identity provider """, allow_none=True).tag(
+        config=True
+    )
+
+    auto_token = Bool(
+        False, help="""Generate token automatically """, allow_none=True
+    ).tag(config=True)
+
+    @default("cookie_secret")
+    def _default_cookie_secret(self):
+        return os.urandom(32)
+
+    authorizer_class = Type(
+        default_value=AllowAllAuthorizer,
+        klass=Authorizer,
+        config=True,
+        help=_("The authorizer class to use."),
+    )
+
+    identity_provider_class = Type(
+        default_value=PasswordIdentityProvider,
+        klass=IdentityProvider,
+        config=True,
+        help=_("The identity provider class to use."),
+    )
+
+    kernel_websocket_connection_class = Type(
+        default_value=ZMQChannelsWebsocketConnection,
+        klass=BaseKernelWebsocketConnection,
+        config=True,
+        help=_("The kernel websocket connection class to use."),
+    )
+
     @property
     def display_url(self):
         if self.custom_display_url:
             url = self.custom_display_url
-            if not url.endswith('/'):
-                url += '/'
+            if not url.endswith("/"):
+                url += "/"
         else:
-            if self.ip in ('', '0.0.0.0'):
-                ip = "%s" % socket.gethostname()
-            else:
-                ip = self.ip
+            ip = "%s" % socket.gethostname() if self.ip in ("", "0.0.0.0") else self.ip
             url = self._url(ip)
         # TODO: do we want to have the token?
-        # if self.token:
-        #     # Don't log full token if it came from config
-        #     token = self.token if self._token_generated else '...'
-        #     url = (url_concat(url, {'token': token})
-        #           + '\n or '
-        #           + url_concat(self._url('127.0.0.1'), {'token': token}))
-        return url
+        if self.identity_provider.token:
+            # Don't log full token if it came from config
+            token = (
+                self.identity_provider.token
+                if self.identity_provider.token_generated
+                else "..."
+            )
+            query = f"?token={token}"
+        else:
+            query = ""
+        return f"{url}{query}"
 
     @property
     def connection_url(self):
-        ip = self.ip if self.ip else 'localhost'
+        ip = self.ip if self.ip else "localhost"
         return self._url(ip)
 
     def _url(self, ip):
         # TODO: https / certfile
         # proto = 'https' if self.certfile else 'http'
-        proto = 'http'
+        proto = "http"
         return "%s://%s:%i%s" % (proto, ip, self.port, self.base_url)
 
     config_file_paths = List(
-        Unicode(),
-        config=True,
-        help=_(
-            'Paths to search for voila.(py|json)'
-        )
+        Unicode(), config=True, help=_("Paths to search for voila.(py|json)")
     )
 
     tornado_settings = Dict(
         {},
         config=True,
         help=_(
-            'Extra settings to apply to tornado application, e.g. headers, ssl, etc'
-        )
+            "Extra settings to apply to tornado application, e.g. headers, ssl, etc"
+        ),
     )
 
-    @default('config_file_paths')
+    @default("config_file_paths")
     def _config_file_paths_default(self):
-        return [os.getcwd()] + jupyter_config_path()
+        return [os.getcwd(), *jupyter_config_path()]
 
-    @default('connection_dir_root')
+    @default("connection_dir_root")
     def _default_connection_dir(self):
         connection_dir = tempfile.gettempdir()
-        self.log.info('Using %s to store connection files' % connection_dir)
+        self.log.info("Using %s to store connection files" % connection_dir)
         return connection_dir
 
-    @default('log_level')
+    @default("log_level")
     def _default_log_level(self):
         return logging.INFO
 
     @property
     def labextensions_path(self):
-        return jupyter_path('labextensions')
+        return jupyter_path("labextensions")
 
-    @default('root_dir')
+    @property
+    def data_dir(self):
+        return get_data_dir()
+
+    @property
+    def schemas_dir(self):
+        return pjoin(self.data_dir, "schemas")
+
+    @property
+    def themes_dir(self):
+        return pjoin(self.data_dir, "themes")
+
+    @default("root_dir")
     def _default_root_dir(self):
         if self.notebook_path:
             return os.path.dirname(os.path.abspath(self.notebook_path))
@@ -354,6 +438,7 @@ class Voila(Application):
         """
         if sys.platform.startswith("win") and sys.version_info >= (3, 8):
             import asyncio
+
             try:
                 from asyncio import (
                     WindowsProactorEventLoopPolicy,
@@ -363,7 +448,10 @@ class Voila(Application):
                 pass
                 # not affected
             else:
-                if type(asyncio.get_event_loop_policy()) is WindowsProactorEventLoopPolicy:
+                if (
+                    type(asyncio.get_event_loop_policy())
+                    is WindowsProactorEventLoopPolicy
+                ):
                     # WindowsProactorEventLoopPolicy is not compatible with tornado 6
                     # fallback to the pre-3.8 default of Selector
                     asyncio.set_event_loop_policy(WindowsSelectorEventLoopPolicy())
@@ -372,7 +460,7 @@ class Voila(Application):
         self._init_asyncio_patch()
         self.log.debug("Searching path %s for config files", self.config_file_paths)
         # to make config_file_paths settable via cmd line, we first need to parse it
-        super(Voila, self).initialize(argv)
+        super().initialize(argv)
         if len(self.extra_args) == 1:
             arg = self.extra_args[0]
             # I am not sure why we need to check if self.notebook_path is set, can we get rid of this?
@@ -382,12 +470,14 @@ class Voila(Application):
                 elif os.path.isfile(arg):
                     self.notebook_path = arg
                 else:
-                    raise ValueError('argument is neither a file nor a directory: %r' % arg)
+                    raise ValueError(
+                        "argument is neither a file nor a directory: %r" % arg
+                    )
         elif len(self.extra_args) != 0:
-            raise ValueError('provided more than 1 argument: %r' % self.extra_args)
+            raise ValueError("provided more than 1 argument: %r" % self.extra_args)
 
         # then we load the config
-        self.load_config_file('voila', path=self.config_file_paths)
+        self.load_config_file("voila", path=self.config_file_paths)
         # common configuration options between the server extension and the application
         self.voila_configuration = VoilaConfiguration(parent=self)
         self.setup_template_dirs()
@@ -396,9 +486,14 @@ class Voila(Application):
     def setup_template_dirs(self):
         if self.voila_configuration.template:
             template_name = self.voila_configuration.template
-            self.template_paths = collect_template_paths(['voila', 'nbconvert'], template_name, prune=True)
-            self.static_paths = collect_static_paths(['voila', 'nbconvert'], template_name)
-            conf_paths = [os.path.join(d, 'conf.json') for d in self.template_paths]
+            self.template_paths = collect_template_paths(
+                ["voila", "nbconvert"], template_name, prune=True
+            )
+            self.static_paths = collect_static_paths(
+                ["voila", "nbconvert"], template_name
+            )
+            self.static_paths.append(DEFAULT_STATIC_FILES_PATH)
+            conf_paths = [os.path.join(d, "conf.json") for d in self.template_paths]
             for p in conf_paths:
                 # see if config file exists
                 if os.path.exists(p):
@@ -406,39 +501,36 @@ class Voila(Application):
                     with open(p) as json_file:
                         conf = json.load(json_file)
                     # update the overall config with it, preserving CLI config priority
-                    if 'traitlet_configuration' in conf:
-                        recursive_update(conf['traitlet_configuration'], self.voila_configuration.config.VoilaConfiguration)
+                    if "traitlet_configuration" in conf:
+                        recursive_update(
+                            conf["traitlet_configuration"],
+                            self.voila_configuration.config.VoilaConfiguration,
+                        )
                         # pass merged config to overall Voilà config
-                        self.voila_configuration.config.VoilaConfiguration = Config(conf['traitlet_configuration'])
-        self.log.debug('using template: %s', self.voila_configuration.template)
-        self.log.debug('template paths:\n\t%s', '\n\t'.join(self.template_paths))
-        self.log.debug('static paths:\n\t%s', '\n\t'.join(self.static_paths))
+                        self.voila_configuration.config.VoilaConfiguration = Config(
+                            conf["traitlet_configuration"]
+                        )
+        self.log.debug("using template: %s", self.voila_configuration.template)
+        self.log.debug("template paths:\n\t%s", "\n\t".join(self.template_paths))
+        self.log.debug("static paths:\n\t%s", "\n\t".join(self.static_paths))
         if self.notebook_path and not os.path.exists(self.notebook_path):
-            raise ValueError('Notebook not found: %s' % self.notebook_path)
+            raise ValueError("Notebook not found: %s" % self.notebook_path)
 
-    def _handle_signal_stop(self, sig, frame):
-        self.log.info('Handle signal %s.' % sig)
-        self.ioloop.add_callback_from_signal(self.ioloop.stop)
-
-    def start(self):
-        self.connection_dir = tempfile.mkdtemp(
-            prefix='voila_',
-            dir=self.connection_dir_root
-        )
-        self.log.info('Storing connection files in %s.' % self.connection_dir)
-        self.log.info('Serving static files from %s.' % self.static_root)
-
+    def init_settings(self) -> Dict:
+        """Initialize settings for Voila application."""
         # default server_url to base_url
         self.server_url = self.server_url or self.base_url
 
-        self.kernel_spec_manager = KernelSpecManager(
-            parent=self
-        )
+        self.kernel_spec_manager = KernelSpecManager(parent=self)
 
         # we create a config manager that load both the serverconfig and nbconfig (classical notebook)
-        read_config_path = [os.path.join(p, 'serverconfig') for p in jupyter_config_path()]
-        read_config_path += [os.path.join(p, 'nbconfig') for p in jupyter_config_path()]
-        self.config_manager = ConfigManager(parent=self, read_config_path=read_config_path)
+        read_config_path = [
+            os.path.join(p, "serverconfig") for p in jupyter_config_path()
+        ]
+        read_config_path += [os.path.join(p, "nbconfig") for p in jupyter_config_path()]
+        self.config_manager = ConfigManager(
+            parent=self, read_config_path=read_config_path
+        )
         self.contents_manager = LargeFileManager(parent=self)
         preheat_kernel: bool = self.voila_configuration.preheat_kernel
         pool_size: int = self.voila_configuration.default_pool_size
@@ -456,21 +548,54 @@ class Voila(Application):
             connection_dir=self.connection_dir,
             kernel_spec_manager=self.kernel_spec_manager,
             allowed_message_types=[
-                'comm_open',
-                'comm_close',
-                'comm_msg',
-                'comm_info_request',
-                'kernel_info_request',
-                'shutdown_request'
-            ]
+                "comm_open",
+                "comm_close",
+                "comm_msg",
+                "comm_info_request",
+                "kernel_info_request",
+                "shutdown_request",
+            ],
         )
 
-        jenv_opt = {"autoescape": True}  # we might want extra options via cmd line like notebook server
-        env = jinja2.Environment(loader=jinja2.FileSystemLoader(self.template_paths), extensions=['jinja2.ext.i18n'], **jenv_opt)
-        nbui = gettext.translation('nbui', localedir=os.path.join(ROOT, 'i18n'), fallback=True)
-        env.install_gettext_translations(nbui, newstyle=False)
+        jenv_opt = {
+            "autoescape": True
+        }  # we might want extra options via cmd line like notebook server
+        env = jinja2.Environment(
+            loader=jinja2.FileSystemLoader(self.template_paths),
+            extensions=["jinja2.ext.i18n"],
+            **jenv_opt,
+        )
+        server_env = jinja2.Environment(
+            loader=jinja2.FileSystemLoader(DEFAULT_TEMPLATE_PATH_LIST),
+            extensions=["jinja2.ext.i18n"],
+            **jenv_opt,
+        )
 
-        self.app = tornado.web.Application(
+        nbui = gettext.translation(
+            "nbui", localedir=os.path.join(ROOT, "i18n"), fallback=True
+        )
+        env.install_gettext_translations(nbui, newstyle=False)
+        server_env.install_gettext_translations(nbui, newstyle=False)
+
+        identity_provider_kwargs = {
+            "parent": self,
+            "log": self.log,
+            "login_handler_class": VoilaLoginHandler,
+        }
+        if self.token is None and not self.auto_token:
+            identity_provider_kwargs["token"] = ""
+        elif self.token is not None:
+            identity_provider_kwargs["token"] = self.token
+
+        self.identity_provider = self.identity_provider_class(
+            **identity_provider_kwargs
+        )
+
+        self.authorizer = self.authorizer_class(
+            parent=self, log=self.log, identity_provider=self.identity_provider
+        )
+
+        settings = dict(
             base_url=self.base_url,
             server_url=self.server_url or self.base_url,
             kernel_manager=self.kernel_manager,
@@ -478,112 +603,160 @@ class Voila(Application):
             allow_remote_access=True,
             autoreload=self.autoreload,
             voila_jinja2_env=env,
-            jinja2_env=env,
-            static_path='/',
-            server_root_dir='/',
+            jinja2_env=server_env,
+            static_path="/",
+            server_root_dir="/",
             contents_manager=self.contents_manager,
-            config_manager=self.config_manager
+            config_manager=self.config_manager,
+            cookie_secret=self.cookie_secret,
+            authorizer=self.authorizer,
+            identity_provider=self.identity_provider,
+            kernel_websocket_connection_class=self.kernel_websocket_connection_class,
+            login_url=url_path_join(self.base_url, "/login"),
         )
+        settings[self.name] = self  # Why???
 
-        self.app.settings.update(self.tornado_settings)
+        return settings
 
+    def init_handlers(self) -> List:
+        """Initialize handlers for Voila application."""
         handlers = []
 
-        handlers.extend([
-            (url_path_join(self.server_url, r'/api/kernels/%s' % _kernel_id_regex), KernelHandler),
-            (url_path_join(self.server_url, r'/api/kernels/%s/channels' % _kernel_id_regex), ZMQChannelsHandler),
-            (
-                url_path_join(self.server_url, r'/voila/templates/(.*)'),
-                TemplateStaticFileHandler
-            ),
-            (
-                url_path_join(self.server_url, r'/voila/static/(.*)'),
-                MultiStaticFileHandler,
-                {
-                    'paths': self.static_paths,
-                    'default_filename': 'index.html'
-                },
-            ),
-            (
-                url_path_join(self.server_url, r'/voila/themes/(.*)'),
-                ThemesHandler,
-                {
-                    'themes_url': '/voila/themes',
-                    'path': '',
-                    'labextensions_path': jupyter_path('labextensions'),
-                    'no_cache_paths': ['/']
-                },
-            ),
-            (url_path_join(self.server_url, r'/voila/api/shutdown/(.*)'), VoilaShutdownKernelHandler)
-        ])
-
-        if preheat_kernel:
+        handlers.extend(
+            [
+                (
+                    url_path_join(
+                        self.server_url, r"/api/kernels/%s" % _kernel_id_regex
+                    ),
+                    KernelHandler,
+                ),
+                (
+                    url_path_join(
+                        self.server_url, r"/api/kernels/%s/channels" % _kernel_id_regex
+                    ),
+                    KernelWebsocketHandler,
+                ),
+                (
+                    url_path_join(self.server_url, r"/voila/templates/(.*)"),
+                    TemplateStaticFileHandler,
+                ),
+                (
+                    url_path_join(self.server_url, r"/voila/static/(.*)"),
+                    MultiStaticFileHandler,
+                    {"paths": self.static_paths, "default_filename": "index.html"},
+                ),
+                (
+                    url_path_join(self.server_url, r"/voila/api/themes/(.*)"),
+                    ThemesHandler,
+                    {
+                        "themes_url": "/voila/api/themes",
+                        "path": self.themes_dir,
+                        "labextensions_path": jupyter_path("labextensions"),
+                        "no_cache_paths": ["/"],
+                    },
+                ),
+                (
+                    url_path_join(self.server_url, r"/voila/api/shutdown/(.*)"),
+                    VoilaShutdownKernelHandler,
+                ),
+            ]
+        )
+        handlers.extend(self.identity_provider.get_handlers())
+        if self.voila_configuration.preheat_kernel:
             handlers.append(
                 (
-                    url_path_join(self.server_url, r'/voila/query/%s' % _kernel_id_regex),
-                    RequestInfoSocketHandler
+                    url_path_join(
+                        self.server_url, r"/voila/query/%s" % _kernel_id_regex
+                    ),
+                    RequestInfoSocketHandler,
                 )
             )
         # Serving JupyterLab extensions
         handlers.append(
             (
-                url_path_join(self.server_url, r'/voila/labextensions/(.*)'),
+                url_path_join(self.server_url, r"/voila/labextensions/(.*)"),
                 FileFindHandler,
                 {
-                    'path': self.labextensions_path,
-                    'no_cache_paths': ['/'],  # don't cache anything in labextensions
+                    "path": self.labextensions_path,
+                    "no_cache_paths": ["/"],  # don't cache anything in labextensions
                 },
             )
         )
         handlers.append(
             (
-                url_path_join(self.server_url, r'/voila/files/(.*)'),
+                url_path_join(self.server_url, r"/voila/files/(.*)"),
                 WhiteListFileHandler,
                 {
-                    'whitelist': self.voila_configuration.file_whitelist,
-                    'blacklist': self.voila_configuration.file_blacklist,
-                    'path': self.root_dir,
+                    "whitelist": self.voila_configuration.file_whitelist,
+                    "blacklist": self.voila_configuration.file_blacklist,
+                    "path": self.root_dir,
                 },
             )
         )
 
-        tree_handler_conf = {
-            'voila_configuration': self.voila_configuration
-        }
+        tree_handler_conf = {"voila_configuration": self.voila_configuration}
         if self.notebook_path:
-            handlers.append((
-                url_path_join(self.server_url, r'/(.*)'),
-                VoilaHandler,
-                {
-                    'notebook_path': os.path.relpath(self.notebook_path, self.root_dir),
-                    'template_paths': self.template_paths,
-                    'config': self.config,
-                    'voila_configuration': self.voila_configuration,
-                    'prelaunch_hook': self.prelaunch_hook
-                }
-            ))
+            handlers.append(
+                (
+                    url_path_join(self.server_url, r"/(.*)"),
+                    TornadoVoilaHandler,
+                    {
+                        "notebook_path": os.path.relpath(
+                            self.notebook_path, self.root_dir
+                        ),
+                        "template_paths": self.template_paths,
+                        "config": self.config,
+                        "voila_configuration": self.voila_configuration,
+                        "prelaunch_hook": self.prelaunch_hook,
+                    },
+                )
+            )
         else:
-            self.log.debug('serving directory: %r', self.root_dir)
-            handlers.extend([
-                (self.server_url, VoilaTreeHandler, tree_handler_conf),
-                (url_path_join(self.server_url, r'/voila/tree' + path_regex),
-                 VoilaTreeHandler, tree_handler_conf),
-                (url_path_join(self.server_url, r'/voila/render/(.*)'),
-                 VoilaHandler,
-                 {
-                     'template_paths': self.template_paths,
-                     'config': self.config,
-                     'voila_configuration': self.voila_configuration,
-                     'prelaunch_hook': self.prelaunch_hook
-                }),
-            ])
+            self.log.debug("serving directory: %r", self.root_dir)
+            handlers.extend(
+                [
+                    (self.server_url, TornadoVoilaTreeHandler, tree_handler_conf),
+                    (
+                        url_path_join(self.server_url, r"/voila/tree" + path_regex),
+                        TornadoVoilaTreeHandler,
+                        tree_handler_conf,
+                    ),
+                    (
+                        url_path_join(self.server_url, r"/voila/render/(.*)"),
+                        TornadoVoilaHandler,
+                        {
+                            "template_paths": self.template_paths,
+                            "config": self.config,
+                            "voila_configuration": self.voila_configuration,
+                            "prelaunch_hook": self.prelaunch_hook,
+                        },
+                    ),
+                ]
+            )
+        return handlers
 
-        self.app.add_handlers('.*$', handlers)
+    def start(self):
+        self.connection_dir = tempfile.mkdtemp(
+            prefix="voila_", dir=self.connection_dir_root
+        )
+        self.log.info("Storing connection files in %s." % self.connection_dir)
+        self.log.info("Serving static files from %s." % self.static_root)
+
+        settings = self.init_settings()
+
+        self.app = tornado.web.Application(**settings)
+        self.app.settings.update(self.tornado_settings)
+        handlers = self.init_handlers()
+        self.app.add_handlers(".*$", handlers)
         self.listen()
+
+    def _handle_signal_stop(self, sig, frame):
+        self.log.info("Handle signal %s." % sig)
+        self.ioloop.add_callback_from_signal(self.ioloop.stop)
 
     def stop(self):
         shutil.rmtree(self.connection_dir)
-        run_sync(self.kernel_manager.shutdown_all())
+        run_sync(self.kernel_manager.shutdown_all)()
 
     def random_ports(self, port, n):
         """Generate a list of n random ports near the given port.
@@ -593,21 +766,26 @@ class Voila(Application):
         """
         for i in range(min(5, n)):
             yield port + i
-        for i in range(n-5):
-            yield max(1, port + random.randint(-2*n, 2*n))
+        for i in range(n - 5):
+            yield max(1, port + random.randint(-2 * n, 2 * n))
 
     def listen(self):
         success = False
-        for port in self.random_ports(self.port, self.port_retries+1):
+        for port in self.random_ports(self.port, self.port_retries + 1):
             try:
                 self.app.listen(port, self.ip)
                 self.port = port
-                self.log.info('Voilà is running at:\n%s' % self.display_url)
-            except socket.error as e:
+                self.log.info("Voilà is running at:\n%s" % self.display_url)
+            except OSError as e:
                 if e.errno == errno.EADDRINUSE:
-                    self.log.info(_('The port %i is already in use, trying another port.') % port)
+                    self.log.info(
+                        _("The port %i is already in use, trying another port.") % port
+                    )
                     continue
-                elif e.errno in (errno.EACCES, getattr(errno, 'WSAEACCES', errno.EACCES)):
+                elif e.errno in (
+                    errno.EACCES,
+                    getattr(errno, "WSAEACCES", errno.EACCES),
+                ):
                     self.log.warning(_("Permission to listen on port %i denied") % port)
                     continue
                 else:
@@ -618,8 +796,12 @@ class Voila(Application):
                 break
 
         if not success:
-            self.log.critical(_('ERROR: the Voilà server could not be started because '
-                              'no available port could be found.'))
+            self.log.critical(
+                _(
+                    "ERROR: the Voilà server could not be started because "
+                    "no available port could be found."
+                )
+            )
             self.exit(1)
 
         if self.open_browser:
@@ -629,7 +811,7 @@ class Voila(Application):
         try:
             self.ioloop.start()
         except KeyboardInterrupt:
-            self.log.info('Stopping...')
+            self.log.info("Stopping...")
         finally:
             self.stop()
 
@@ -637,33 +819,38 @@ class Voila(Application):
         try:
             browser = webbrowser.get(self.browser or None)
         except webbrowser.Error as e:
-            self.log.warning(_('No web browser found: %s.') % e)
+            self.log.warning(_("No web browser found: %s.") % e)
             browser = None
 
         if not browser:
             return
 
         uri = self.base_url
-        fd, open_file = tempfile.mkstemp(suffix='.html')
+        fd, open_file = tempfile.mkstemp(suffix=".html")
         # Write a temporary file to open in the browser
-        with io.open(fd, 'w', encoding='utf-8') as fh:
-            # TODO: do we want to have the token?
-            # if self.token:
-            #     url = url_concat(url, {'token': self.token})
+        with open(fd, "w", encoding="utf-8") as fh:
             url = url_path_join(self.connection_url, uri)
 
-            include_assets_functions = create_include_assets_functions(self.voila_configuration.template, url)
+            include_assets_functions = create_include_assets_functions(
+                self.voila_configuration.template, url
+            )
 
-            jinja2_env = self.app.settings['jinja2_env']
-            template = jinja2_env.get_template('browser-open.html')
-            fh.write(template.render(
-                open_url=url, base_url=url,
-                theme=self.voila_configuration.theme,
-                **include_assets_functions
-            ))
+            jinja2_env = self.app.settings["voila_jinja2_env"]
+            template = jinja2_env.get_template("browser-open.html")
+            fh.write(
+                template.render(
+                    open_url=url,
+                    base_url=url,
+                    theme=self.voila_configuration.theme,
+                    **include_assets_functions,
+                )
+            )
 
         def target():
-            return browser.open(urljoin('file:', pathname2url(open_file)), new=self.webbrowser_open_new)
+            return browser.open(
+                urljoin("file:", pathname2url(open_file)), new=self.webbrowser_open_new
+            )
+
         threading.Thread(target=target).start()
 
 
