@@ -8,7 +8,6 @@
  ****************************************************************************/
 
 export * from './app';
-export * from './manager';
 export * from './shell';
-export * from './output';
-export * from './plugins';
+export * from './voilaplugins';
+export * from './tools';
