@@ -1,4 +1,5 @@
 import os
+
 import pytest
 
 MAX_TIMEOUT_SECONDS = 240
@@ -11,16 +12,18 @@ def preheat_mode():
 
 @pytest.fixture
 def voila_args_extra():
-    return [f'--VoilaExecutor.timeout={MAX_TIMEOUT_SECONDS}']
+    return [f"--VoilaExecutor.timeout={MAX_TIMEOUT_SECONDS}"]
 
 
 @pytest.fixture
 def voila_notebook(notebook_directory):
-    return os.path.join(notebook_directory, 'many_iopub_messages.ipynb')
+    return os.path.join(notebook_directory, "many_iopub_messages.ipynb")
 
 
 @pytest.mark.filterwarnings("ignore")
 async def test_template_cwd(http_server_client, base_url):
-    response = await http_server_client.fetch(base_url, request_timeout=MAX_TIMEOUT_SECONDS+20)
-    html_text = response.body.decode('utf-8')
-    assert 'you should see me' in html_text
+    response = await http_server_client.fetch(
+        base_url, request_timeout=MAX_TIMEOUT_SECONDS + 20
+    )
+    html_text = response.body.decode("utf-8")
+    assert "you should see me" in html_text
