@@ -1,6 +1,6 @@
 import pytest
 
-NOTEBOOK_PATH = 'no_kernelspec.ipynb'
+NOTEBOOK_PATH = "no_kernelspec.ipynb"
 
 
 @pytest.fixture
@@ -10,13 +10,13 @@ def non_existing_kernel_notebook(base_url):
 
 @pytest.fixture
 def voila_args(notebook_directory, voila_args_extra):
-    return ['--VoilaTest.root_dir=%r' % notebook_directory] + voila_args_extra
+    return ["--VoilaTest.root_dir=%r" % notebook_directory, *voila_args_extra]
 
 
-async def test_non_existing_kernel(http_server_client,
-                                   non_existing_kernel_notebook,
-                                   ):
-
+async def test_non_existing_kernel(
+    http_server_client,
+    non_existing_kernel_notebook,
+):
     response = await http_server_client.fetch(non_existing_kernel_notebook)
     assert response.code == 200
-    assert 'Executing without a kernelspec' in response.body.decode('utf-8')
+    assert "Executing without a kernelspec" in response.body.decode("utf-8")
