@@ -1,4 +1,5 @@
 import pytest
+
 NOTEBOOK_PATH = "non_existing_kernel.ipynb"
 
 
@@ -9,13 +10,10 @@ def non_existing_kernel_notebook(base_url):
 
 @pytest.fixture
 def voila_args(notebook_directory, voila_args_extra):
-    return ["--VoilaTest.root_dir=%r" % notebook_directory] + voila_args_extra
+    return ["--VoilaTest.root_dir=%r" % notebook_directory, *voila_args_extra]
 
 
-async def test_non_existing_kernel(
-    http_server_client, non_existing_kernel_notebook
-):
-
+async def test_non_existing_kernel(http_server_client, non_existing_kernel_notebook):
     response = await http_server_client.fetch(non_existing_kernel_notebook)
     assert response.code == 200
     assert "non-existing kernel" in response.body.decode("utf-8")
