@@ -1,7 +1,8 @@
-import pytest
-import time
 import asyncio
 import os
+import time
+
+import pytest
 
 
 @pytest.fixture
@@ -11,7 +12,7 @@ def preheat_mode():
 
 @pytest.fixture
 def voila_notebook(notebook_directory):
-    return os.path.join(notebook_directory, 'preheat', 'pre_heat.ipynb')
+    return os.path.join(notebook_directory, "preheat", "pre_heat.ipynb")
 
 
 NOTEBOOK_EXECUTION_TIME = 2
@@ -23,7 +24,7 @@ async def send_request(sc, url, wait=0):
     real_time = time.time()
     response = await sc.fetch(url)
     real_time = time.time() - real_time
-    html_text = response.body.decode('utf-8')
+    html_text = response.body.decode("utf-8")
     return real_time, html_text
 
 
@@ -34,7 +35,7 @@ async def test_request_before_kernel_heated(http_server_client, base_url):
     to display result.
     """
     time, text = await send_request(sc=http_server_client, url=base_url)
-    assert 'hello world' in text
+    assert "hello world" in text
     assert time > NOTEBOOK_EXECUTION_TIME
     await asyncio.sleep(NOTEBOOK_EXECUTION_TIME + 1)
 
@@ -44,29 +45,27 @@ async def test_render_time_with_preheated_kernel(http_server_client, base_url):
     We wait for kernel properly heated before sending request,
     it should take under `TIME_THRESHOLD` second to return result
     """
-    time, text = await send_request(sc=http_server_client,
-                                    url=base_url,
-                                    wait=NOTEBOOK_EXECUTION_TIME + 2)
-    assert 'hello world' in text
+    time, text = await send_request(
+        sc=http_server_client, url=base_url, wait=NOTEBOOK_EXECUTION_TIME + 2
+    )
+    assert "hello world" in text
     assert time < TIME_THRESHOLD
     await asyncio.sleep(NOTEBOOK_EXECUTION_TIME + 1)
 
 
-async def test_render_time_with_multiple_requests(http_server_client,
-                                                  base_url):
+async def test_render_time_with_multiple_requests(http_server_client, base_url):
     """
     We send a second request just after the first one, so the
     pool is not filled yet and a normal kernel is used instead
     """
     time_list = []
     for wait in [NOTEBOOK_EXECUTION_TIME + 1, 0]:
-
-        time, _ = await send_request(sc=http_server_client,
-                                     url=base_url,
-                                     wait=wait)
+        time, _ = await send_request(sc=http_server_client, url=base_url, wait=wait)
         time_list.append(time)
 
-    assert time_list[1] > time_list[0]  # Render time for a normal kernel is bigger than a a preheated kernel
+    assert (
+        time_list[1] > time_list[0]
+    )  # Render time for a normal kernel is bigger than a a preheated kernel
     await asyncio.sleep(NOTEBOOK_EXECUTION_TIME + 1)
 
 
@@ -75,10 +74,10 @@ async def test_request_with_query(http_server_client, base_url):
     We sent request with query parameter, preheat kernel should
     be activated.
     """
-    url = f'{base_url}?foo=bar'
-    time, _ = await send_request(sc=http_server_client,
-                                 url=url,
-                                 wait=NOTEBOOK_EXECUTION_TIME + 1)
+    url = f"{base_url}?foo=bar"
+    time, _ = await send_request(
+        sc=http_server_client, url=url, wait=NOTEBOOK_EXECUTION_TIME + 1
+    )
     assert time < TIME_THRESHOLD
     await asyncio.sleep(NOTEBOOK_EXECUTION_TIME + 1)
 
@@ -90,11 +89,11 @@ async def test_request_with_theme_parameter(http_server_client, base_url):
     """
     wait = NOTEBOOK_EXECUTION_TIME + 2
 
-    url = f'{base_url}?voila-theme=dark'
+    url = f"{base_url}?theme=dark"
     time, _ = await send_request(sc=http_server_client, url=url, wait=wait)
     assert time > TIME_THRESHOLD
 
-    url = f'{base_url}?voila-theme=light'
+    url = f"{base_url}?theme=light"
     time, _ = await send_request(sc=http_server_client, url=url, wait=wait)
     assert time < TIME_THRESHOLD
 
@@ -106,6 +105,6 @@ async def test_request_with_template_parameter(http_server_client, base_url):
     """
     wait = NOTEBOOK_EXECUTION_TIME + 2
 
-    url = f'{base_url}?voila-template=lab'
+    url = f"{base_url}?template=lab"
     time, _ = await send_request(sc=http_server_client, url=url, wait=wait)
     assert time < TIME_THRESHOLD
