@@ -16,7 +16,7 @@ def preheat_mode():
 
 @pytest.fixture
 def voila_notebook(notebook_directory):
-    return os.path.join(notebook_directory, 'preheat', 'default_env_variables.ipynb')
+    return os.path.join(notebook_directory, "preheat", "default_env_variables.ipynb")
 
 
 NOTEBOOK_EXECUTION_TIME = 2
@@ -25,11 +25,11 @@ NOTEBOOK_EXECUTION_TIME = 2
 async def send_request(sc, url, wait=0):
     await asyncio.sleep(wait)
     response = await sc.fetch(url)
-    return response.body.decode('utf-8')
+    return response.body.decode("utf-8")
 
 
 async def test_default_kernel_env_variable(http_server_client, base_url):
-    html_text = await send_request(sc=http_server_client,
-                                   url=base_url,
-                                   wait=NOTEBOOK_EXECUTION_TIME + 1)
-    assert 'BAR' in html_text
+    html_text = await send_request(
+        sc=http_server_client, url=base_url, wait=NOTEBOOK_EXECUTION_TIME + 1
+    )
+    assert "BAR" in html_text
