@@ -1,7 +1,8 @@
-import pytest
-import time
 import asyncio
 import os
+import time
+
+import pytest
 
 BASE_DIR = os.path.dirname(__file__)
 NOTEBOOK_EXECUTION_TIME = 3
@@ -11,8 +12,8 @@ TIME_THRESHOLD = 1
 
 @pytest.fixture
 def voila_config_file_paths_arg():
-    path = os.path.join(BASE_DIR, '..', 'configs', 'preheat')
-    return '--VoilaTest.config_file_paths=[%r]' % path
+    path = os.path.join(BASE_DIR, "..", "configs", "preheat")
+    return "--VoilaTest.config_file_paths=[%r]" % path
 
 
 @pytest.fixture
@@ -22,7 +23,7 @@ def preheat_mode():
 
 @pytest.fixture
 def voila_notebook(notebook_directory):
-    return os.path.join(notebook_directory, 'preheat')
+    return os.path.join(notebook_directory, "preheat")
 
 
 async def send_request(sc, url, wait=0):
@@ -35,18 +36,24 @@ async def send_request(sc, url, wait=0):
 
 
 async def test_render_notebook_with_heated_kernel(http_server_client, base_url):
-    await asyncio.sleep(NUMBER_PREHEATED_KERNEL*NOTEBOOK_EXECUTION_TIME + 1)
-    time, text = await send_request(sc=http_server_client, url=f'{base_url}voila/render/pre_heat.ipynb')
+    await asyncio.sleep(NUMBER_PREHEATED_KERNEL * NOTEBOOK_EXECUTION_TIME + 1)
+    time, text = await send_request(
+        sc=http_server_client, url=f"{base_url}voila/render/pre_heat.ipynb"
+    )
 
-    assert 'hello world' in text
+    assert "hello world" in text
     assert time < TIME_THRESHOLD
     await asyncio.sleep(NOTEBOOK_EXECUTION_TIME + 1)
 
 
-async def test_render_blacklisted_notebook_with_nornal_kernel(http_server_client, base_url):
-    await asyncio.sleep(NUMBER_PREHEATED_KERNEL*NOTEBOOK_EXECUTION_TIME + 1)
-    time, text = await send_request(sc=http_server_client, url=f'{base_url}voila/render/blacklisted.ipynb')
+async def test_render_blacklisted_notebook_with_nornal_kernel(
+    http_server_client, base_url
+):
+    await asyncio.sleep(NUMBER_PREHEATED_KERNEL * NOTEBOOK_EXECUTION_TIME + 1)
+    time, text = await send_request(
+        sc=http_server_client, url=f"{base_url}voila/render/blacklisted.ipynb"
+    )
 
-    assert 'hello world' in text
+    assert "hello world" in text
     assert time > TIME_THRESHOLD
     await asyncio.sleep(NOTEBOOK_EXECUTION_TIME + 1)
