@@ -1,9 +1,9 @@
 # test serving a notebook or python/c++ notebook
 import os
+
 import pytest
 
-
-TEST_XEUS_CLING = os.environ.get('VOILA_TEST_XEUS_CLING', '') == '1'
+TEST_XEUS_CLING = os.environ.get("VOILA_TEST_XEUS_CLING", "") == "1"
 
 
 @pytest.fixture
@@ -13,14 +13,14 @@ def preheat_mode():
 
 @pytest.fixture
 def voila_args(notebook_directory, voila_args_extra):
-    return ['--VoilaTest.root_dir=%r' % notebook_directory] + voila_args_extra
+    return ["--VoilaTest.root_dir=%r" % notebook_directory, *voila_args_extra]
 
 
 async def test_print(http_server_client, print_notebook_url):
     print(print_notebook_url)
     response = await http_server_client.fetch(print_notebook_url)
     assert response.code == 200
-    assert 'Hi Voilà' in response.body.decode('utf-8')
+    assert "Hi Voilà" in response.body.decode("utf-8")
 
 
 @pytest.fixture
@@ -30,14 +30,18 @@ def voila_args_extra():
 
 async def test_print_py(http_server_client, print_notebook_url):
     print(print_notebook_url)
-    response = await http_server_client.fetch(print_notebook_url.replace('ipynb', 'py'))
+    response = await http_server_client.fetch(print_notebook_url.replace("ipynb", "py"))
     assert response.code == 200
-    assert 'Hi Voilà' in response.body.decode('utf-8')
+    assert "Hi Voilà" in response.body.decode("utf-8")
 
 
-@pytest.mark.skipif(not TEST_XEUS_CLING, reason='opt in to avoid having to install xeus-cling')
+@pytest.mark.skipif(
+    not TEST_XEUS_CLING, reason="opt in to avoid having to install xeus-cling"
+)
 async def test_print_julia_notebook(http_server_client, print_notebook_url):
     print(print_notebook_url)
-    response = await http_server_client.fetch(print_notebook_url.replace('.ipynb', '_cpp.ipynb'))
+    response = await http_server_client.fetch(
+        print_notebook_url.replace(".ipynb", "_cpp.ipynb")
+    )
     assert response.code == 200
-    assert 'Hi Voilà, from c++' in response.body.decode('utf-8')
+    assert "Hi Voilà, from c++" in response.body.decode("utf-8")
