@@ -1,8 +1,8 @@
-import pytest
-
 import os
 
-NOTEBOOK_PATH = 'syntax_error.ipynb'
+import pytest
+
+NOTEBOOK_PATH = "syntax_error.ipynb"
 
 
 @pytest.fixture
@@ -11,9 +11,8 @@ def voila_notebook(notebook_directory):
 
 
 async def test_syntax_error(http_server_client, base_url):
-
     response = await http_server_client.fetch(base_url)
     assert response.code == 200
-    output = response.body.decode('utf-8')
-    assert 'There was an error when executing cell' in output
-    assert 'This should not be executed' not in output
+    output = response.body.decode("utf-8")
+    assert "There was an error when executing cell" in output
+    assert "This should not be executed" not in output
