@@ -13,7 +13,7 @@ test.describe('Voila performance Tests', () => {
   });
   test('Render tree classic', async ({ page, browserName }, testInfo) => {
     const testFunction = async () => {
-      await page.goto('?voila-template=classic');
+      await page.goto('?template=classic');
       // wait for page to load
       await page.waitForSelector('.list-header');
     };
@@ -45,7 +45,7 @@ test.describe('Voila performance Tests', () => {
 
   test('Render tree dark theme', async ({ page, browserName }, testInfo) => {
     const testFunction = async () => {
-      await page.goto('?voila-theme=dark');
+      await page.goto('?theme=dark');
       // wait for page to load
       await page.waitForSelector('.list-header');
     };
@@ -61,7 +61,7 @@ test.describe('Voila performance Tests', () => {
 
   test('Render tree miami theme', async ({ page, browserName }, testInfo) => {
     const testFunction = async () => {
-      await page.goto('?voila-theme=jupyterlab_miami_nights');
+      await page.goto('?theme=JupyterLab%20Miami%20Nights');
       // wait for page to load
       await page.waitForSelector('.list-header');
     };
@@ -81,9 +81,7 @@ test.describe('Voila performance Tests', () => {
   }, testInfo) => {
     const notebookName = 'basics';
     const testFunction = async () => {
-      await page.goto(
-        `/voila/render/${notebookName}.ipynb?voila-template=classic`
-      );
+      await page.goto(`/voila/render/${notebookName}.ipynb?template=classic`);
       // wait for the widgets to load
       await page.waitForSelector('span[role="presentation"] >> text=x');
     };
@@ -105,14 +103,15 @@ test.describe('Voila performance Tests', () => {
     const testFunction = async () => {
       await page.goto(`/voila/render/${notebookName}.ipynb`);
       // wait for the widgets to load
-      await page.waitForSelector('span[role="presentation"] >> text=x');
+      await page.waitForSelector('span.mjx-char >> text=x');
     };
     await addBenchmarkToTest(notebookName, testFunction, testInfo, browserName);
     // change the value of the slider
     await page.fill('text=4.00', '8.00');
     await page.keyboard.down('Enter');
+    await page.waitForTimeout(500);
     // fetch the value of the label
-    const value = await page.$eval('input', el => el.value);
+    const value = await page.$eval('input', (el) => el.value);
 
     expect(value).toBe('64');
     // wait for the final MathJax message to be hidden
@@ -127,7 +126,7 @@ test.describe('Voila performance Tests', () => {
   }, testInfo) => {
     const notebookName = 'basics';
     const testFunction = async () => {
-      await page.goto(`/voila/render/${notebookName}.ipynb?voila-theme=dark`);
+      await page.goto(`/voila/render/${notebookName}.ipynb?theme=dark`);
       // wait for the widgets to load
       await page.waitForSelector('span[role="presentation"] >> text=x');
     };
@@ -146,7 +145,7 @@ test.describe('Voila performance Tests', () => {
     const notebookName = 'basics';
     const testFunction = async () => {
       await page.goto(
-        `/voila/render/${notebookName}.ipynb?voila-theme=jupyterlab_miami_nights`
+        `/voila/render/${notebookName}.ipynb?theme=JupyterLab%20Miami%20Nights`
       );
       // wait for the widgets to load
       await page.waitForSelector('span[role="presentation"] >> text=x');
@@ -169,14 +168,14 @@ test.describe('Voila performance Tests', () => {
   });
 
   test('Render 404 error with classic template', async ({ page }) => {
-    await page.goto('/voila/render/unknown.ipynb?voila-template=classic');
+    await page.goto('/voila/render/unknown.ipynb?template=classic');
     await page.waitForSelector('.voila-error');
 
     expect(await page.screenshot()).toMatchSnapshot('404-classic.png');
   });
 
   test('Render 404 error with dark theme', async ({ page }) => {
-    await page.goto('/voila/render/unknown.ipynb?voila-theme=dark');
+    await page.goto('/voila/render/unknown.ipynb?theme=dark');
     await page.waitForSelector('.voila-error');
 
     expect(await page.screenshot()).toMatchSnapshot('404-dark.png');
@@ -257,6 +256,34 @@ test.describe('Voila performance Tests', () => {
     expect(await page.screenshot()).toMatchSnapshot(`${notebookName}.png`);
   });
 
+  test('Render and benchmark mimerenderers.ipynb', async ({
+    page,
+    browserName
+  }, testInfo) => {
+    const notebookName = 'mimerenderers';
+    const testFunction = async () => {
+      await page.goto(`/voila/render/${notebookName}.ipynb`);
+      await page.waitForSelector('.biojs_msa_div');
+      await page.waitForSelector('#MathJax_Message', { state: 'hidden' });
+      await page.waitForTimeout(2000);
+    };
+    await addBenchmarkToTest(notebookName, testFunction, testInfo, browserName);
+    expect(await page.screenshot()).toMatchSnapshot(`${notebookName}.png`);
+  });
+
+  test('Render and benchmark bokeh.ipynb', async ({
+    page,
+    browserName
+  }, testInfo) => {
+    const notebookName = 'bokeh';
+    const testFunction = async () => {
+      await page.goto(`/voila/render/${notebookName}.ipynb`);
+      await page.waitForSelector('.bk-Canvas');
+    };
+    await addBenchmarkToTest(notebookName, testFunction, testInfo, browserName);
+    expect(await page.screenshot()).toMatchSnapshot(`${notebookName}.png`);
+  });
+
   test('Benchmark the multiple widgets notebook', async ({
     page,
     browserName
@@ -276,6 +303,7 @@ test.describe('Voila performance Tests', () => {
     );
     expect(await page.screenshot()).toMatchSnapshot(`${notebookName}.png`);
   });
+
   test('Render and benchmark query-strings.ipynb', async ({
     page,
     browserName
@@ -298,6 +326,7 @@ test.describe('Voila performance Tests', () => {
     await page.waitForSelector('#MathJax_Message', { state: 'hidden' });
     expect(await page.screenshot()).toMatchSnapshot(`${notebookName}.png`);
   });
+
   test('Render and benchmark reveal.ipynb', async ({
     page,
     browserName
@@ -308,6 +337,13 @@ test.describe('Voila performance Tests', () => {
       await page.waitForSelector('span[role="presentation"] >> text=x');
     };
     await addBenchmarkToTest(notebookName, testFunction, testInfo, browserName);
+    expect(await page.screenshot()).toMatchSnapshot(`${notebookName}.png`);
+  });
+
+  test('Render yaml.ipynb', async ({ page, browserName }, testInfo) => {
+    const notebookName = 'yaml';
+    await page.goto(`/voila/render/${notebookName}.ipynb`);
+    await page.waitForSelector('span >> text=hey');
     expect(await page.screenshot()).toMatchSnapshot(`${notebookName}.png`);
   });
 });
