@@ -8,13 +8,18 @@
 #############################################################################
 
 from ._version import __version__  # noqa
-from .server_extension import _load_jupyter_server_extension # noqa
+from .server_extension import _load_jupyter_server_extension  # noqa
 from .server_extension import load_jupyter_server_extension  # noqa
 
 
 def _jupyter_nbextension_paths():
     return [
-        dict(section="notebook", src="static", dest="voila", require="voila/extension")
+        {
+            "section": "notebook",
+            "src": "static",
+            "dest": "voila",
+            "require": "voila/extension",
+        }
     ]
 
 
