@@ -11,11 +11,10 @@
 # Distributed under the terms of the Modified BSD License.
 
 import re
-
 from collections import namedtuple
 
 # Use "hatch version xx.yy.zz" to handle version changes
-__version__ = "0.5.0a1"
+__version__ = "0.5.0a5"
 
 # PEP440 version parser
 _version_regex = re.compile(
@@ -33,17 +32,16 @@ _version_regex = re.compile(
 
 _version_fields = _version_regex.match(__version__).groupdict()  # type:ignore
 
-VersionInfo = namedtuple("VersionInfo", ["major", "minor", "micro", "releaselevel", "serial"])
+VersionInfo = namedtuple(
+    "VersionInfo", ["major", "minor", "micro", "releaselevel", "serial"]
+)
 
 version_info = VersionInfo(
     *[
-        field
-        for field in (
-            int(_version_fields["major"]),
-            int(_version_fields["minor"]),
-            int(_version_fields["micro"]),
-            _version_fields["releaselevel"] or "",
-            _version_fields["serial"] or "",
-        )
+        int(_version_fields["major"]),
+        int(_version_fields["minor"]),
+        int(_version_fields["micro"]),
+        _version_fields["releaselevel"] or "",
+        _version_fields["serial"] or "",
     ]
 )
