@@ -7,31 +7,52 @@
 # The full license is in the file LICENSE, distributed with this software.  #
 #############################################################################
 
-import os
 import json
+import os
 
-from jupyter_core.paths import jupyter_path
 import nbconvert.exporters.templateexporter
-
+from jupyter_core.paths import jupyter_path
 
 ROOT = os.path.dirname(__file__)
-STATIC_ROOT = os.path.join(ROOT, 'static')
+STATIC_ROOT = os.path.join(ROOT, "static")
 # if the directory above us contains the following paths, it means we are installed in dev mode (pip install -e .)
-DEV_MODE = os.path.exists(os.path.join(ROOT, '../setup.py')) and os.path.exists(os.path.join(ROOT, '../share'))
+DEV_MODE = os.path.exists(os.path.join(ROOT, "../setup.py")) and os.path.exists(
+    os.path.join(ROOT, "../share")
+)
 
 
-def collect_template_paths(app_names, template_name='default', prune=False, root_dirs=None):
-    return collect_paths(app_names, template_name, include_root_paths=True, prune=prune, root_dirs=root_dirs)
-
-
-def collect_static_paths(app_names, template_name='default', prune=False, root_dirs=None):
+def collect_template_paths(
+    app_names, template_name="default", prune=False, root_dirs=None
+):
     return collect_paths(
-        app_names, template_name, include_root_paths=False, prune=prune, root_dirs=root_dirs, subdir='static'
+        app_names,
+        template_name,
+        include_root_paths=True,
+        prune=prune,
+        root_dirs=root_dirs,
+    )
+
+
+def collect_static_paths(
+    app_names, template_name="default", prune=False, root_dirs=None
+):
+    return collect_paths(
+        app_names,
+        template_name,
+        include_root_paths=False,
+        prune=prune,
+        root_dirs=root_dirs,
+        subdir="static",
     )
 
 
 def collect_paths(
-    app_names, template_name='default', subdir=None, include_root_paths=True, prune=False, root_dirs=None
+    app_names,
+    template_name="default",
+    subdir=None,
+    include_root_paths=True,
+    prune=False,
+    root_dirs=None,
 ):
     """
     Voilà supports custom templates for rendering notebooks.
@@ -61,7 +82,7 @@ def collect_paths(
     for template_name in template_names:
         for root_dir in root_dirs:
             for app_name in app_names:
-                app_dir = os.path.join(root_dir, app_name, 'templates')
+                app_dir = os.path.join(root_dir, app_name, "templates")
                 path = os.path.join(app_dir, template_name)
                 if subdir:
                     path = os.path.join(path, subdir)
@@ -80,7 +101,7 @@ def collect_paths(
             # {% extends 'nbconvert/templates/classic/base.html' %}
             paths.append(root_dir)
             for app_name in app_names:
-                app_dir = os.path.join(root_dir, app_name, 'templates')
+                app_dir = os.path.join(root_dir, app_name, "templates")
                 # we include app_dir for when we want to be explicit, but less than root_dir, e.g.
                 # {% extends 'classic/base.html' %}
                 paths.append(app_dir)
@@ -88,7 +109,9 @@ def collect_paths(
     if not found_at_least_one:
         paths = "\n\t".join(full_paths)
         raise ValueError(
-            'No template sub-directory with name %r found in the following paths:\n\t%s' % (template_name, paths)
+            "No template sub-directory with name {!r} found in the following paths:\n\t{}".format(
+                template_name, paths
+            )
         )
     return paths
 
@@ -98,9 +121,19 @@ def _default_root_dirs():
     # relative to the package directory (first entry, meaning with highest precedence)
     root_dirs = []
     if DEV_MODE:
-        root_dirs.append(os.path.abspath(os.path.join(ROOT, '..', 'share', 'jupyter')))
+        root_dirs.append(os.path.abspath(os.path.join(ROOT, "..", "share", "jupyter")))
     if nbconvert.exporters.templateexporter.DEV_MODE:
-        root_dirs.append(os.path.abspath(os.path.join(nbconvert.exporters.templateexporter.ROOT, '..', '..', 'share', 'jupyter')))
+        root_dirs.append(
+            os.path.abspath(
+                os.path.join(
+                    nbconvert.exporters.templateexporter.ROOT,
+                    "..",
+                    "..",
+                    "share",
+                    "jupyter",
+                )
+            )
+        )
     root_dirs.extend(jupyter_path())
 
     return root_dirs
@@ -113,19 +146,21 @@ def _find_template_hierarchy(app_names, template_name, root_dirs):
         conf = {}
         for root_dir in root_dirs:
             for app_name in app_names:
-                conf_file = os.path.join(root_dir, app_name, 'templates', template_name, 'conf.json')
+                conf_file = os.path.join(
+                    root_dir, app_name, "templates", template_name, "conf.json"
+                )
                 if os.path.exists(conf_file):
                     with open(conf_file) as f:
                         new_conf = json.load(f)
                         new_conf.update(conf)
                         conf = new_conf
-        if 'base_template' in conf:
-            template_name = conf['base_template']
+        if "base_template" in conf:
+            template_name = conf["base_template"]
         else:
-            if template_name == 'base':
+            if template_name == "base":
                 # the default template has no base_template
                 template_name = None
             else:
                 # if not specified, 'base' is assumed
-                template_name = 'base'
+                template_name = "base"
     return template_names
